fix: puleendo backend

+++ src/presentation/sorteo/controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { prisma } from "../../data";
+import { isValidObjectId } from "../../helpers";
 
 export class SorteoController{
     constructor(){
@@ -18,7 +19,7 @@ export class SorteoController{
               }
           });
         
-          return res.status(200).json({ sorteoData });
+          return res.status(200).json({sorteoData});
         
 
     }
@@ -26,9 +27,10 @@ export class SorteoController{
     async getSorteosOne(req:Request, res:Response){
 
         const {id}=req.params;
-        if(!id) return res.json(400).json({ error:'no enviaste un id' });
-
         try {
+            const isMongoId=isValidObjectId(id);            
+            if(!isMongoId) return res.status(400).json({ error:'no enviaste un id valido' });
+
             const sorteoGetOne = await prisma.sorteo.findUnique({
             where: {
                 id: id,
@@ -43,18 +45,20 @@ export class SorteoController{
     }
 
     async postSorteos(req:Request, res:Response){
-        const {name,description,startDate,image, winner , status} = req.body;
-
-        const sorteoData={
-            name,
-            description,
-            startDate,
-            image,
-            winner,
-            status
-        }
+        const {name,description,startDate,image, winner , status,usuarioId} = req.body;
 
         try {
+            const sorteoData={
+                name,
+                usuarioId,
+                description,
+                startDate,
+                image,
+                winner,
+                status,
+            }
+            if (!sorteoData.usuarioId) return res.status(400).json({error:'el sorteo debe ser creado por un usuario existente'});
+
             const sorteoPost = await prisma.sorteo.create({
                 data:sorteoData
             });
@@ -62,7 +66,7 @@ export class SorteoController{
             return res.status(200).json({ post:true,sorteoPost });
         } catch (error) {
             console.log(error);
-            return res.status(400).json({error:'Error al crear el sorteo'});
+            return res.status(500).json({error:'Error al crear el sorteo'});
         }
 
     }
@@ -71,6 +75,8 @@ export class SorteoController{
 
         const {id}=req.params;
         if(!id) return res.status(400).json({ error:'no enviaste un id' });
+        const isMongoId=isValidObjectId(id);            
+        if(!isMongoId) return res.status(400).json({ error:'no enviaste un id valido' });
 
         const {name,description,startDate,status,image,winner}=req.body;
         const sorteoBody={
@@ -103,17 +109,18 @@ export class SorteoController{
 
         const {id}=req.params;
         if(!id) return res.status(400).json({ error:'no enviaste un id' });
+        const isMongoId=isValidObjectId(id);            
+        if(!isMongoId) return res.status(400).json({ error:'no enviaste un id valido' });
 
         try {
 
-
-            const participantDelete=await prisma.participante.deleteMany({
+            const findId=await prisma.sorteo.findUnique({
                 where:{
-                    sorteoId:id
+                    id:id
                 }
             })
-
-
+            if(!findId) return res.status(400).json({ error:'no se encontro al usuario a eliminar en la bd' });
+            
             const sorteoDelete=await prisma.sorteo.delete({
                 where:{
                     id:id
